backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-openai
langchain-groq
langchain-community
langchain-text-splitters
openai
groq
pinecone
opik
requests
beautifulsoup4
tiktoken
httpx
